Fix compiler warnings about narrowing conversion

+++ src/mapping/EKFSlam/EKFSlam.cpp
@@ -168,14 +168,14 @@ std::vector<ObstaclePoint> EKFSlam::getObstacles() {
   // is there a difference between () and [] indexing for these matrices?
   // lemme look it seems like either works....
   for (int i = 2 * obstacles.size() + 3; i < mu.rows() - 1; i = i + 2) {
-    ObstaclePoint ob{mu(i), mu(i+1)};
+    ObstaclePoint ob{static_cast<float>(mu(i)), static_cast<float>(mu(i+1))};
     obstacles.push_back(ob);
   }
   return obstacles;
 }
 
 PointXY EKFSlam::getPosition() {
-  PointXY location{mu[0], mu[1]};
+	PointXY location{static_cast<float>(mu[0]), static_cast<float>(mu[1])};
   return location;
 }
